harness: split package doc, H, and Suite into different files
The size of testing.go was a bit unwieldy. Split sub_test.go to match.
