+++ harness/harness.go
@@ -13,136 +13,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package harness provides a reusable test framework akin to the standard
-// "testing" Go package. For now there is no automated code generation
-// component like the "go test" command but that may be a future extension.
-// Test functions must be of type `func(*harness.H)` and registered directly
-// with a test Suite struct which can then be launched via the Run method.
-//
-// Within these functions, use the Error, Fail or related methods to signal failure.
-//
-// Tests may be skipped if not applicable with a call to
-// the Skip method of *H:
-//     func NeedsSomeData(h *harness.H) {
-//         if os.Getenv("SOME_DATA") == "" {
-//             h.Skip("skipping test due to missing SOME_DATA")
-//         }
-//         ...
-//     }
-//
-// Subtests
-//
-// The Run method of H allow defining subtests,
-// without having to define separate functions for each. This enables uses
-// like table-driven and hierarchical tests.
-// It also provides a way to share common setup and tear-down code:
-//
-//     func Foo(h *harness.H) {
-//         // <setup code>
-//         h.Run("A=1", func(h *harness.H) { ... })
-//         h.Run("A=2", func(h *harness.H) { ... })
-//         h.Run("B=1", func(h *harness.H) { ... })
-//         // <tear-down code>
-//     }
-//
-// Each subtest has a unique name: the combination of the name
-// of the top-level test and the sequence of names passed to Run, separated by
-// slashes, with an optional trailing sequence number for disambiguation.
-//
-// The argument to the -harness.run command-line flag is an unanchored regular
-// expression that matches the test's name. For tests with multiple slash-separated
-// elements, such as subtests, the argument is itself slash-separated, with
-// expressions matching each name element in turn. Because it is unanchored, an
-// empty expression matches any string.
-// For example, using "matching" to mean "whose name contains":
-//
-//     go run foo.go -harness.run ''      # Run all tests.
-//     go run foo.go -harness.run Foo     # Run top-level tests matching "Foo", such as "TestFooBar".
-//     go run foo.go -harness.run Foo/A=  # For top-level tests matching "Foo", run subtests matching "A=".
-//     go run foo.go -harness.run /A=1    # For all top-level tests, run subtests matching "A=1".
-//
-// Subtests can also be used to control parallelism. A parent test will only
-// complete once all of its subtests complete. In this example, all tests are
-// run in parallel with each other, and only with each other, regardless of
-// other top-level tests that may be defined:
-//
-//     func GroupedParallel(h *harness.H) {
-//         for _, tc := range tests {
-//             tc := tc // capture range variable
-//             h.Run(tc.Name, func(h *harness.H) {
-//                 h.Parallel()
-//                 ...
-//             })
-//         }
-//     }
-//
-// Run does not return until parallel subtests have completed, providing a way
-// to clean up after a group of parallel tests:
-//
-//     func TeardownParallel(h *harness.H) {
-//         // This Run will not return until the parallel tests finish.
-//         h.Run("group", func(h *harness.H) {
-//             h.Run("Test1", parallelTest1)
-//             h.Run("Test2", parallelTest2)
-//             h.Run("Test3", parallelTest3)
-//         })
-//         // <tear-down code>
-//     }
-//
-// Suite
-//
-// Individual tests are grouped into a test suite in order to execute them.
-// TODO: this part of the API deviates from the "testing" package and is TBD.
-//
-// A simple implementation of a test suite:
-//
-//	func SomeTest(h *harness.H) {
-//		h.Skip("TODO")
-//	}
-//
-//	func main() {
-//		suite := harness.NewSuite([]InternalTest{
-//			{"SomeTest", SomeTest},
-//		})
-//		os.Exit(suite.Run())
-//	}
-//
 package harness
 
 import (
 	"bytes"
 	"context"
-	"flag"
 	"fmt"
 	"io"
 	"log"
-	"os"
 	"runtime"
-	"runtime/debug"
-	"runtime/pprof"
-	"runtime/trace"
 	"sync"
 	"time"
-)
-
-var (
-	// The directory in which to create profile files and the like. When run from
-	// "go test", the binary always runs in the source directory for the package;
-	// this flag lets "go test" tell the binary to write the files in the directory where
-	// the "go test" command is run.
-	outputDir = flag.String("harness.outputdir", "", "write profiles to `dir`")
-
-	// Report as tests are run; default is silent for success.
-	chatty           = flag.Bool("harness.v", false, "verbose: print additional output")
-	match            = flag.String("harness.run", "", "run only tests matching `regexp`")
-	memProfile       = flag.String("harness.memprofile", "", "write a memory profile to `file`")
-	memProfileRate   = flag.Int("harness.memprofilerate", 0, "set memory profiling `rate` (see runtime.MemProfileRate)")
-	cpuProfile       = flag.String("harness.cpuprofile", "", "write a cpu profile to `file`")
-	blockProfile     = flag.String("harness.blockprofile", "", "write a goroutine blocking profile to `file`")
-	blockProfileRate = flag.Int("harness.blockprofilerate", 1, "set blocking profile `rate` (see runtime.SetBlockProfileRate)")
-	traceFile        = flag.String("harness.trace", "", "write an execution trace to `file`")
-	timeout          = flag.Duration("harness.timeout", 0, "fail test binary execution after duration `d` (0 means unlimited)")
-	parallel         = flag.Int("harness.parallel", runtime.GOMAXPROCS(0), "run at most `n` tests in parallel")
 )
 
 // H is a type passed to Test functions to manage test state and support formatted test logs.
@@ -413,13 +294,6 @@ func (t *H) Parallel() {
 	t.start = time.Now()
 }
 
-// An internal type but exported because it is cross-package; part of the implementation
-// of the "go test" command.
-type InternalTest struct {
-	Name string
-	F    func(*H)
-}
-
 func tRunner(t *H, fn func(t *H)) {
 	t.ctx, t.cancel = context.WithCancel(t.parentContext())
 	defer t.cancel()
@@ -514,93 +388,6 @@ func (t *H) Run(name string, f func(t *H)) bool {
 	return !t.failed
 }
 
-// Suite is a type passed to a TestMain function to run the actual tests.
-// Suite manages the execution of a set of test functions.
-type Suite struct {
-	tests  []InternalTest
-	match  *matcher
-	chatty bool
-
-	// mu protects the following fields which are used to manage
-	// parallel test execution.
-	mu sync.Mutex
-
-	// Channel used to signal tests that are ready to be run in parallel.
-	startParallel chan bool
-
-	// running is the number of tests currently running in parallel.
-	// This does not include tests that are waiting for subtests to complete.
-	running int
-
-	// numWaiting is the number tests waiting to be run in parallel.
-	numWaiting int
-
-	// maxParallel is a copy of the parallel flag.
-	maxParallel int
-}
-
-func (c *Suite) waitParallel() {
-	c.mu.Lock()
-	if c.running < c.maxParallel {
-		c.running++
-		c.mu.Unlock()
-		return
-	}
-	c.numWaiting++
-	c.mu.Unlock()
-	<-c.startParallel
-}
-
-func (c *Suite) release() {
-	c.mu.Lock()
-	if c.numWaiting == 0 {
-		c.running--
-		c.mu.Unlock()
-		return
-	}
-	c.numWaiting--
-	c.mu.Unlock()
-	c.startParallel <- true // Pick a waiting test to be run.
-}
-
-// NewSuite
-func NewSuite(tests []InternalTest) *Suite {
-	return &Suite{
-		tests:         tests,
-		startParallel: make(chan bool),
-	}
-}
-
-// Run runs the tests. It returns an exit code to pass to os.Exit.
-func (s *Suite) Run() int {
-	// The user may have already called flag.Parse.
-	if !flag.Parsed() {
-		flag.Parse()
-	}
-
-	// Initialize Suite based on command line flags.
-	// TODO(marineam): offer other ways to do this.
-	s.match = newMatcher(*match, "-harness.run")
-	s.maxParallel = *parallel
-	s.chatty = *chatty
-	s.running = 1 // Set the count to 1 for the main (sequential) test.
-	s.before()
-	startAlarm()
-	testRan, testOk := s.runTests()
-	if !testRan {
-		fmt.Fprintln(os.Stderr, "testing: warning: no tests to run")
-	}
-	if !testOk {
-		fmt.Println("FAIL")
-		s.after()
-		return 1
-	}
-
-	fmt.Println("PASS")
-	s.after()
-	return 0
-}
-
 func (t *H) report() {
 	if t.parent == nil {
 		return
@@ -615,142 +402,5 @@ func (t *H) report() {
 		} else {
 			t.flushToParent(format, "PASS", t.name, dstr)
 		}
-	}
-}
-
-func (s *Suite) runTests() (ran, ok bool) {
-	t := &H{
-		signal:  make(chan bool),
-		barrier: make(chan bool),
-		w:       os.Stdout,
-		suite:   s,
-	}
-	tRunner(t, func(t *H) {
-		for _, test := range s.tests {
-			t.Run(test.Name, test.F)
-		}
-		// Run catching the signal rather than the tRunner as a separate
-		// goroutine to avoid adding a goroutine during the sequential
-		// phase as this pollutes the stacktrace output when aborting.
-		go func() { <-t.signal }()
-	})
-	return t.ran, !t.Failed()
-}
-
-// before runs before all testing.
-func (m *Suite) before() {
-	if *memProfileRate > 0 {
-		runtime.MemProfileRate = *memProfileRate
-	}
-	if *cpuProfile != "" {
-		f, err := os.Create(toOutputDir(*cpuProfile))
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "testing: %s", err)
-			return
-		}
-		if err := pprof.StartCPUProfile(f); err != nil {
-			fmt.Fprintf(os.Stderr, "testing: can't start cpu profile: %s", err)
-			f.Close()
-			return
-		}
-		// Could save f so after can call f.Close; not worth the effort.
-	}
-	if *traceFile != "" {
-		f, err := os.Create(toOutputDir(*traceFile))
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "testing: %s", err)
-			return
-		}
-		if err := trace.Start(f); err != nil {
-			fmt.Fprintf(os.Stderr, "testing: can't start tracing: %s", err)
-			f.Close()
-			return
-		}
-		// Could save f so after can call f.Close; not worth the effort.
-	}
-	if *blockProfile != "" && *blockProfileRate >= 0 {
-		runtime.SetBlockProfileRate(*blockProfileRate)
-	}
-}
-
-// after runs after all testing.
-func (m *Suite) after() {
-	if *cpuProfile != "" {
-		pprof.StopCPUProfile() // flushes profile to disk
-	}
-	if *traceFile != "" {
-		trace.Stop() // flushes trace to disk
-	}
-	if *memProfile != "" {
-		f, err := os.Create(toOutputDir(*memProfile))
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "testing: %s\n", err)
-			os.Exit(2)
-		}
-		runtime.GC() // materialize all statistics
-		if err = pprof.WriteHeapProfile(f); err != nil {
-			fmt.Fprintf(os.Stderr, "testing: can't write %s: %s\n", *memProfile, err)
-			os.Exit(2)
-		}
-		f.Close()
-	}
-	if *blockProfile != "" && *blockProfileRate >= 0 {
-		f, err := os.Create(toOutputDir(*blockProfile))
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "testing: %s\n", err)
-			os.Exit(2)
-		}
-		if err = pprof.Lookup("block").WriteTo(f, 0); err != nil {
-			fmt.Fprintf(os.Stderr, "testing: can't write %s: %s\n", *blockProfile, err)
-			os.Exit(2)
-		}
-		f.Close()
-	}
-}
-
-// toOutputDir returns the file name relocated, if required, to outputDir.
-// Simple implementation to avoid pulling in path/filepath.
-func toOutputDir(path string) string {
-	if *outputDir == "" || path == "" {
-		return path
-	}
-	if runtime.GOOS == "windows" {
-		// On Windows, it's clumsy, but we can be almost always correct
-		// by just looking for a drive letter and a colon.
-		// Absolute paths always have a drive letter (ignoring UNC).
-		// Problem: if path == "C:A" and outputdir == "C:\Go" it's unclear
-		// what to do, but even then path/filepath doesn't help.
-		// TODO: Worth doing better? Probably not, because we're here only
-		// under the management of go test.
-		if len(path) >= 2 {
-			letter, colon := path[0], path[1]
-			if ('a' <= letter && letter <= 'z' || 'A' <= letter && letter <= 'Z') && colon == ':' {
-				// If path starts with a drive letter we're stuck with it regardless.
-				return path
-			}
-		}
-	}
-	if os.IsPathSeparator(path[0]) {
-		return path
-	}
-	return fmt.Sprintf("%s%c%s", *outputDir, os.PathSeparator, path)
-}
-
-var timer *time.Timer
-
-// startAlarm starts an alarm if requested.
-func startAlarm() {
-	if *timeout > 0 {
-		timer = time.AfterFunc(*timeout, func() {
-			debug.SetTraceback("all")
-			panic(fmt.Sprintf("test timed out after %v", *timeout))
-		})
-	}
-}
-
-// stopAlarm turns off the alarm.
-func stopAlarm() {
-	if *timeout > 0 {
-		timer.Stop()
 	}
 }
